Refactors DeputyshipBuilderResult to extend BuilderResult
