+++ api/app/src/v2/Registration/DeputyshipProcessing/DeputyshipBuilderResult.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\v2\Registration\DeputyshipProcessing;
 
-use App\Model\DeputyshipProcessingRawDbAccessResult;
 use App\v2\Registration\Enum\DeputyshipBuilderResultOutcome;
 use App\v2\Registration\Enum\DeputyshipCandidateAction;
 
@@ -13,84 +12,25 @@ use App\v2\Registration\Enum\DeputyshipCandidateAction;
  * This is the result of building entities for one court order, its associations to deputies and reports,
  * and updates to the order status and deputy status on order (if applicable).
  */
-class DeputyshipBuilderResult
+class DeputyshipBuilderResult extends BuilderResult
 {
-    private int $numCandidatesApplied = 0;
+    public function __construct(protected readonly \UnitEnum $outcome)
+    {
+        if (!$outcome instanceof DeputyshipBuilderResultOutcome) {
+            throw new \TypeError('Incorrect enum type provided. Must provide DeputyshipBuilderResultOutcome');
+        }
 
-    private int $numCandidatesFailed = 0;
-
-    public function __construct(
-        private readonly DeputyshipBuilderResultOutcome $outcome,
-        /** @var string[] $errors */
-        private array $errors = [],
-        /** @var array<string, int> $candidatesApplied */
-        private array $candidatesApplied = [],
-    ) {
+        parent::__construct($outcome);
         // initialise counts of candidates applied
         foreach (DeputyshipCandidateAction::cases() as $case) {
             $this->candidatesApplied[$case->value] = 0;
         }
     }
 
-    public function hasErrors(): bool
+    public function addActionResult(\UnitEnum $actionResult): void
     {
-        return count($this->errors) > 0;
-    }
-
-    public function getOutcome(): DeputyshipBuilderResultOutcome
-    {
-        return $this->outcome;
-    }
-
-    public function getMessage(): string
-    {
-        $message = 'Builder result: failed candidates = ' . $this->numCandidatesFailed .
-            '; applied candidates = ' . $this->numCandidatesApplied;
-
-        $candidateDetails = [];
-        foreach ($this->candidatesApplied as $action => $num) {
-            $candidateDetails[] = $action . ':' . $num;
-        }
-
-        if (count($candidateDetails) > 0) {
-            $message .= '; candidate details = ' . implode('|', $candidateDetails);
-        }
-
-        return $message;
-    }
-
-    public function getErrorMessage(): ?string
-    {
-        if (0 == count($this->errors)) {
-            return null;
-        }
-
-        return 'ERRORS while applying candidates: ' . implode(' / ', $this->errors);
-    }
-
-    /**
-     * @return string[]
-     */
-    public function getErrors(): array
-    {
-        return $this->errors;
-    }
-
-    /**
-     * Record the result of applying a candidate to the database.
-     * If $result->success is true, the insert or update was a success; otherwise, it failed, and exception message is stored.
-     */
-    public function addCandidateResult(DeputyshipProcessingRawDbAccessResult $result): void
-    {
-        if ($result->success) {
-            ++$this->numCandidatesApplied;
-            ++$this->candidatesApplied[$result->action->value];
-        } else {
-            ++$this->numCandidatesFailed;
-
-            if (!is_null($result->error)) {
-                $this->errors[] = $result->error;
-            }
+        if ($actionResult instanceof DeputyshipBuilderResultOutcome) {
+            ++$this->candidatesApplied[$actionResult->value];
         }
     }
 }
